udf_sentimentAnalysis/requirements.txt
scikit-learn
bert-score
python-dotenv
groq
